+++ src/model/solarflareModel.go
@@ -61,7 +61,7 @@ const (
 // StatusResponse returns information on the service and the plugin network to the consumer
 type StatusResponse struct {
 	RegisteredServerCount int     `json:"registeredServerCount"`
-	ClockSpeedBPM         int     `json:"clockSpeedBpm"`
+	ClockSpeedBPM         float64 `json:"clockSpeedBpm"`
 	ClockSpeedMultiplier  float64 `json:"clockSpeedMultiplier"`
 }
